Move email authorization tests to their own file
